+++ p_caritas/src/Componets/document.js
@@ -1,154 +1,84 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React, { Component } from 'react'
-import html2canvas from 'html2canvas';
 import './ficha_usuarios.css';
 import Button from '@material-ui/core/Button';
-import * as jsPDF from 'jspdf';
-import axios from 'axios';
-
-let cant=0;
-let vpsicologica='';
-let vfisica='';
-let veconomica='';
-let vsexual='';
+import { generarPdf } from './Forms/actions';
+let cant = 0;
+let vpsicologica = '';
+let vfisica = '';
+let veconomica = '';
+let vsexual = '';
 let vurbana = '';
 let vrural = '';
 let causaEconomica = '';
 let causaAlcohol = '';
 let causaInfiel = '';
-let victima= '';
+let victima = '';
 let agresor = '';
 
-const port = 'https://apicaritas.herokuapp.com/';
-
-function createCase(params) {
-    var case_number = params.NumeroEx;
-    var patient_id = params.NumeroIdent;
-    var remission_id = params.Remision;
-    
-    let violence_type_id = 0;
-    if (params.VPsicologica)
-        violence_type_id = 1;
-    else if (params.VFisica)
-        violence_type_id = 2;
-    else if (params.VEconomica)
-        violence_type_id = 3;
-    else if (params.VSexual)
-        violence_type_id = 4;
-    
-    var condition_id = params.Victima ? 1 : 2;
-
-    let cause_id = 0;
-    if (params.CEconomica)
-        cause_id = 1;
-    else if(params.CInfidelidad)
-        cause_id = 2;
-    else if (params.CAlcoholismo)
-        cause_id = 3;
-    
-    var location_violence_id = params.VUrbana ? 1 : 2;   
-
-    var mun = params.Parroquia;
-    let city_id = 0;
-    switch (mun) {
-        case 'sps':
-            city_id = 1;
-            break;
-        case 'lima':
-            city_id = 2;
-            break;
-        default:
-            city_id = 4;
-    }
-
-    var state_id = params.EstadoAtencion;
-    var benefitted_amount = Number(params.Ninos) + Number(params.Ninas) + Number(params.Otros);
-        axios.post(port+'/api/casos/'+case_number+'/'+patient_id+'/'+
-        remission_id+'/'+violence_type_id+'/'+condition_id+'/'+cause_id+'/'+location_violence_id+
-        '/'+city_id+'/'+state_id+'/'+benefitted_amount).catch(e=>{console.log(e)}) 
-    
-}
-
 export class Form3 extends Component {
- 
-   back=e=>{
+
+    back = e => {
         e.preventDefault();
         this.props.prevStep();
     }
 
-    generarPdf = () =>{
-        createCase(this.props.vals);
-        
-        const input  = document.getElementById("imprimir");
-        html2canvas(input).then((canvas) => {
-            const imgData = canvas.toDataURL('img/png');
-            const pdf = new jsPDF('p', 'mm', 'legal');
-            pdf.addImage(imgData, 'PNG', 0, 0);
-            pdf.save('Ficha_Paciente.pdf'); 
-        });
-    }
-    componentWillMount=()=>{
-        let ninos,ninas,otros;
-        if(this.props.vals.VPsicologica){
-            vpsicologica=" Psicologica ";
+    componentWillMount = () => {
+        let ninos, ninas, otros;
+        if (this.props.vals.VPsicologica) {
+            vpsicologica = " Psicologica ";
         }
-        if(this.props.vals.VEconomica){
-            veconomica=" Economica ";
-        }        
-        if(this.props.vals.VFisica){
-            vfisica="Fisica";
+        if (this.props.vals.VEconomica) {
+            veconomica = " Economica ";
         }
-        if(this.props.vals.VSexual){
-            vsexual="Sexual";
+        if (this.props.vals.VFisica) {
+            vfisica = "Fisica";
         }
-        if(this.props.vals.Ninos==='')
-            { ninos = 0;}
-        else    
-            {ninos = parseInt(this.props.vals.Ninos);}
-        if(this.props.vals.Ninas==='')
-        {  ninas =0;}
-        else 
-            { ninas = parseInt(this.props.vals.Ninas);}
-        if(this.props.vals.Otros==='')
-            { otros = 0;}    
-        else
-            { otros = parseInt(this.props.vals.Otros);}
-        if(this.props.vals.VUrbana){
+        if (this.props.vals.VSexual) {
+            vsexual = "Sexual";
+        }
+        if (this.props.vals.Ninos === '') { ninos = 0; }
+        else { ninos = parseInt(this.props.vals.Ninos); }
+        if (this.props.vals.Ninas === '') { ninas = 0; }
+        else { ninas = parseInt(this.props.vals.Ninas); }
+        if (this.props.vals.Otros === '') { otros = 0; }
+        else { otros = parseInt(this.props.vals.Otros); }
+        if (this.props.vals.VUrbana) {
             vurbana = "Urbana";
         }
-        if(this.props.vals.VRural)
+        if (this.props.vals.VRural)
             vrural = "Rural";
         if (this.props.vals.CEconomica)
             causaEconomica = "Economica";
-        if(this.props.vals.CAlcoholismo)
+        if (this.props.vals.CAlcoholismo)
             causaAlcohol = 'Alcholimos';
-        if(this.props.vals.CInfidelidad)
+        if (this.props.vals.CInfidelidad)
             causaInfiel = 'Infidelidad';
-        if(this.props.vals.Victima)
+        if (this.props.vals.Victima)
             victima = 'Victima';
-        if(this.props.vals.Agresor)
+        if (this.props.vals.Agresor)
             agresor = 'Agresor';
         cant = ninos + ninas + otros;
 
-        console.log("cantidad es: " + cant + 'Ninos '+otros);
-    }
-    
-    render() {
-        const {vals}=this.props;
 
+    }
+
+    render() {
+        const { vals } = this.props;
         function getDate() {
             const today = new Date(),
-            date = today.getDate() + '-' + (today.getMonth() + 1) + '-' + today.getFullYear();
+                date = today.getDate() + '-' + (today.getMonth() + 1) + '-' + today.getFullYear();
             return date;
         }
         const fechaActual = getDate();
 
         return (
             <div>
-                <div className = "g34"> 
-                    <Button color="secondary" variant="outlined" onClick={this.back}>Regresar</Button><Button  color="primary" variant="outlined" onClick={this.generarPdf}>Imprimir</Button>
+                <div className="right">
+                    <Button color="secondary" variant="outlined" onClick={this.back}>Regresar</Button>
+                    <Button color="primary" variant="outlined" onClick={() => generarPdf(this.props)}>Imprimir</Button>
                 </div>
-                <div id = "imprimir" className="c26">
+                <div id="imprimir" className="c26">
                     <p className="c10"><span className="c3">CARITAS DE SAN PEDRO SULA</span></p>
                     <p className="c10 c19"><span className="c11"></span></p>
                     <p className="c24"><span className="c14">SOLICITUD DE ASISTENCIA PSICOL&Oacute;GICA GRATUITA EN CASOS DE VIOLENCIA DOM&Eacute;STICA</span></p>
@@ -157,7 +87,7 @@ export class Form3 extends Component {
                     <p className="c5 c19"><span className="c17"></span></p>
                     <p className="c5"><span className="c11">1. Datos Personales&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Fecha de recepci&oacute;n: {fechaActual}</span></p>
                     <p className="c5"><span className="c11">A) Beneficiario (a)</span></p>
-                    
+
                     <table className="c16">
                         <tbody>
                             <tr className="c9">
@@ -311,10 +241,10 @@ export class Form3 extends Component {
                             </tr>
                             <tr className="c25">
                                 <td className="c12" colSpan="1" rowSpan="1">
-                                    <input className = "inputText" type="text" name="Importe" />
+                                    <input className="inputText" type="text" name="Importe" />
                                 </td>
                                 <td className="c12" colSpan="1" rowSpan="1">
-                                    <input className = "inputText" type="text" name="Importe" />
+                                    <input className="inputText" type="text" name="Importe" />
                                 </td>
                             </tr>
                         </tbody>
@@ -351,7 +281,7 @@ export class Form3 extends Component {
                     <p className="c10"><span className="c17">(Firma)</span></p>
                 </div>
             </div>
-        
+
         );
     }
 }
